Add `skip_cfa_ranges` option to config.yml
This allows configuring the function analyzer to skip certain
problematic ranges in the program. Only for exceptional cases
that prevents analysis from finishing.

When using this feature, ensure that no `function` symbols
exist for the affected range in symbols.txt, otherwise CFA
will run for the defined functions, regardless of this setting.